tvl_ocr/requirements.txt
torch==2.4.0
transformers==4.37.2
timm
einops
fastapi
uvicorn
pillow
